Move fake_memcached.FakeAsyncMemcached to tests/mocks.py
